changed inMemoryWatchlist to persist data
